Complete the Cats vs Dogs data gathering notebook
This was adding to the existing notebook all the steps of building the
folders, getting the images initialy, splitting them into train
categories etc.
I have also added some checks to ensure that if the prep is completed
already, then running the notebook does not have an effect, like
copying another 1000 images to the valid folder.